+++ src/pages/Groups.tsx
@@ -1,18 +1,13 @@
-import { useContext, useState, useEffect, useMemo } from 'react';
+import { useContext, useState, useEffect } from 'react';
 import { useNavigate } from 'react-router';
 import AuthContext from '@/contexts/auth/authContext';
 import {
   Typography,
   Box,
   Card,
-  CardContent,
-  CardActions,
   Button,
-  Chip,
   CircularProgress,
   Alert,
-  IconButton,
-  Tooltip,
   Snackbar,
   Dialog,
   DialogTitle,
@@ -21,47 +16,27 @@ import {
   DialogActions,
   useMediaQuery,
   useTheme,
-  TextField,
-  InputAdornment,
-  Tabs,
-  Tab,
-  FormControl,
-  InputLabel,
-  Select,
-  MenuItem,
-  Paper,
   SelectChangeEvent,
 } from '@mui/material';
 import AddIcon from '@mui/icons-material/Add';
 import GroupIcon from '@mui/icons-material/Group';
-import ContentCopyIcon from '@mui/icons-material/ContentCopy';
-import EditIcon from '@mui/icons-material/Edit';
-import DeleteIcon from '@mui/icons-material/Delete';
 import PeopleIcon from '@mui/icons-material/People';
-import ExpandMoreIcon from '@mui/icons-material/ExpandMore';
-import ExpandLessIcon from '@mui/icons-material/ExpandLess';
-import ManageAccountsIcon from '@mui/icons-material/ManageAccounts';
-import SearchIcon from '@mui/icons-material/Search';
-import FilterListIcon from '@mui/icons-material/FilterList';
+import DeleteIcon from '@mui/icons-material/Delete';
 import GroupFormDialog from '@/components/GroupFormDialog';
 import JoinGroupDialog from '@/components/JoinGroupDialog';
 import GroupMembersDialog from '@/components/GroupMembersDialog';
+import GroupFilters from '@/components/GroupFilters';
+import GroupCard from '@/components/GroupCard';
 import { groupService, GroupMemberRole } from '@/services/groupService';
 import { Group } from '@/types/Group';
-import { useGroupPermissions, GroupPermission } from '@/hooks/useGroupPermissions';
+import { useGroupPermissions } from '@/hooks/useGroupPermissions';
+import { useGroupFilters, SortOption } from '@/hooks/useGroupFilters';
 import { copyToClipboard } from '@/utils/clipboard';
-
-// Constants
-const DESCRIPTION_MAX_LENGTH = 100;
 
 interface Notification {
   message: string;
   type: 'success' | 'error' | 'info';
 }
-
-// Filter types
-type FilterTab = 'all' | 'admin' | 'moderator' | 'member';
-type SortOption = 'name' | 'members' | 'recent';
 
 export default function Groups() {
   const auth = useContext(AuthContext);
@@ -69,11 +44,17 @@ export default function Groups() {
   const { hasPermission, getRoleColor, getRoleLabel } = useGroupPermissions();
   const theme = useTheme();
   const isMobile = useMediaQuery(theme.breakpoints.down('sm'));
+
+  // Dialog states
   const [isDialogOpen, setIsDialogOpen] = useState(false);
   const [isJoinDialogOpen, setIsJoinDialogOpen] = useState(false);
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
   const [dialogMode, setDialogMode] = useState<'create' | 'edit'>('create');
   const [selectedGroup, setSelectedGroup] = useState<Group | null>(null);
+  const [isMembersDialogOpen, setIsMembersDialogOpen] = useState(false);
+  const [selectedGroupForMembers, setSelectedGroupForMembers] = useState<Group | null>(null);
+
+  // Data states
   const [groups, setGroups] = useState<Group[]>([]);
   const [memberCounts, setMemberCounts] = useState<Record<string, number>>({});
   const [expandedDescriptions, setExpandedDescriptions] = useState<Record<string, boolean>>({});
@@ -82,14 +63,20 @@ export default function Groups() {
   const [notification, setNotification] = useState<Notification | null>(null);
   const [isDeleting, setIsDeleting] = useState(false);
   const [deletingGroupId, setDeletingGroupId] = useState<string | null>(null);
-  const [isMembersDialogOpen, setIsMembersDialogOpen] = useState(false);
-  const [selectedGroupForMembers, setSelectedGroupForMembers] = useState<Group | null>(null);
 
-  // Filter states
-  const [searchQuery, setSearchQuery] = useState('');
-  const [activeFilterTab, setActiveFilterTab] = useState<FilterTab>('all');
-  const [sortOption, setSortOption] = useState<SortOption>('name');
-  const [showFilters, setShowFilters] = useState(false);
+  // Use our custom hook for filtering
+  const {
+    searchQuery,
+    activeFilterTab,
+    sortOption,
+    showFilters,
+    filteredGroups,
+    handleSearchChange,
+    handleFilterTabChange,
+    handleSortChange,
+    toggleFilters,
+    clearFilters,
+  } = useGroupFilters({ groups, memberCounts });
 
   // Set up real-time subscription for user groups
   useEffect(() => {
@@ -179,67 +166,7 @@ export default function Groups() {
     };
   }, [auth?.myUser?.userId]);
 
-  // Filter and sort groups
-  const filteredGroups = useMemo(() => {
-    let result = [...groups];
-
-    // Apply search filter
-    if (searchQuery.trim()) {
-      const query = searchQuery.toLowerCase().trim();
-      result = result.filter(
-        group =>
-          group.groupName.toLowerCase().includes(query) ||
-          group.groupDescription.toLowerCase().includes(query) ||
-          group.joinCode.toLowerCase().includes(query)
-      );
-    }
-
-    // Apply role filter
-    if (activeFilterTab !== 'all') {
-      result = result.filter(group => {
-        switch (activeFilterTab) {
-          case 'admin':
-            return group.userRole === GroupMemberRole.ADMIN;
-          case 'moderator':
-            return group.userRole === GroupMemberRole.MODERATOR;
-          case 'member':
-            return group.userRole === GroupMemberRole.MEMBER;
-          default:
-            return true;
-        }
-      });
-    }
-
-    // Apply sorting
-    result.sort((a, b) => {
-      switch (sortOption) {
-        case 'name':
-          return a.groupName.localeCompare(b.groupName);
-        case 'members':
-          return (memberCounts[b.groupId] || 0) - (memberCounts[a.groupId] || 0);
-        case 'recent': {
-          // Handle Timestamp type properly
-          // Using a more specific approach for Firestore Timestamp
-          const getTimestamp = (timestamp: unknown): number => {
-            if (!timestamp) return 0;
-            // Check if it's a Firestore Timestamp with toMillis method
-            if (typeof timestamp === 'object' && timestamp !== null && 'toMillis' in timestamp) {
-              return (timestamp as { toMillis: () => number }).toMillis();
-            }
-            // Fallback to timestamp value if it's a number
-            return typeof timestamp === 'number' ? timestamp : 0;
-          };
-
-          return getTimestamp(b.createdAt) - getTimestamp(a.createdAt);
-        }
-        default:
-          return 0;
-      }
-    });
-
-    return result;
-  }, [groups, searchQuery, activeFilterTab, sortOption, memberCounts]);
-
+  // Dialog handlers
   const handleOpenCreateDialog = () => {
     setDialogMode('create');
     setSelectedGroup(null);
@@ -292,16 +219,6 @@ export default function Groups() {
       ...prev,
       [groupId]: !prev[groupId],
     }));
-  };
-
-  const truncateDescription = (description: string, groupId: string) => {
-    if (description.length <= DESCRIPTION_MAX_LENGTH) {
-      return description;
-    }
-
-    return expandedDescriptions[groupId]
-      ? description
-      : `${description.substring(0, DESCRIPTION_MAX_LENGTH)}...`;
   };
 
   const handleOpenJoinDialog = () => {
@@ -396,193 +313,6 @@ export default function Groups() {
     setSelectedGroupForMembers(null);
   };
 
-  // Filter handlers
-  const handleSearchChange = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setSearchQuery(event.target.value);
-  };
-
-  const handleFilterTabChange = (event: React.SyntheticEvent, newValue: FilterTab) => {
-    setActiveFilterTab(newValue);
-  };
-
-  const handleSortChange = (event: SelectChangeEvent<SortOption>) => {
-    setSortOption(event.target.value as SortOption);
-  };
-
-  const toggleFilters = () => {
-    setShowFilters(!showFilters);
-  };
-
-  const renderGroupCard = (group: Group) => {
-    const isDeletingThisGroup = deletingGroupId === group.groupId;
-    const isAdmin = group.userRole === GroupMemberRole.ADMIN;
-    const isModerator = group.userRole === GroupMemberRole.MODERATOR;
-    const canEdit = isAdmin || isModerator;
-    const canDelete = isAdmin;
-    const memberCount = memberCounts[group.groupId] || 0;
-
-    return (
-      <Card
-        key={group.groupId}
-        sx={{
-          height: '100%',
-          display: 'flex',
-          flexDirection: 'column',
-          cursor: isDeletingThisGroup ? 'not-allowed' : 'pointer',
-          transition: 'transform 0.2s, box-shadow 0.2s',
-          opacity: isDeletingThisGroup ? 0.7 : 1,
-          '&:hover': {
-            transform: isDeletingThisGroup ? 'none' : 'translateY(-4px)',
-            boxShadow: isDeletingThisGroup ? 1 : 4,
-          },
-        }}
-        onClick={() => !isDeletingThisGroup && handleNavigateToGroupDetail(group.groupId)}
-      >
-        <CardContent sx={{ flexGrow: 1, p: { xs: 2, sm: 3 } }}>
-          <Box sx={{ display: 'flex', alignItems: 'center', mb: { xs: 1.5, sm: 2 } }}>
-            <GroupIcon
-              sx={{ mr: 1, color: 'primary.main', fontSize: { xs: '1.2rem', sm: '1.5rem' } }}
-            />
-            <Typography
-              variant="h6"
-              component="h3"
-              sx={{
-                flexGrow: 1,
-                fontSize: { xs: '1rem', sm: '1.25rem' },
-              }}
-            >
-              {group.groupName}
-              {isDeletingThisGroup && (
-                <CircularProgress size={16} sx={{ ml: 1, display: 'inline-block' }} />
-              )}
-            </Typography>
-            {canEdit && !isDeletingThisGroup && (
-              <Tooltip title="Edit Group">
-                <IconButton
-                  size="small"
-                  onClick={e => {
-                    e.stopPropagation(); // Prevent card click when clicking edit button
-                    handleOpenEditDialog(group);
-                  }}
-                  sx={{ ml: 1 }}
-                >
-                  <EditIcon fontSize="small" />
-                </IconButton>
-              </Tooltip>
-            )}
-            {canDelete && !isDeletingThisGroup && (
-              <Tooltip title="Delete Group">
-                <IconButton
-                  size="small"
-                  onClick={e => {
-                    e.stopPropagation(); // Prevent card click when clicking delete button
-                    handleOpenDeleteDialog(group);
-                  }}
-                  sx={{ ml: 1 }}
-                  color="error"
-                >
-                  <DeleteIcon fontSize="small" />
-                </IconButton>
-              </Tooltip>
-            )}
-          </Box>
-          <Box sx={{ mb: { xs: 1.5, sm: 2 } }}>
-            <Typography
-              variant="body2"
-              color="text.secondary"
-              sx={{ fontSize: { xs: '0.8rem', sm: '0.875rem' } }}
-            >
-              {truncateDescription(group.groupDescription, group.groupId)}
-            </Typography>
-            {group.groupDescription.length > DESCRIPTION_MAX_LENGTH && (
-              <Button
-                size="small"
-                onClick={e => {
-                  e.stopPropagation(); // Prevent card click when clicking show more/less
-                  toggleDescription(group.groupId);
-                }}
-                sx={{
-                  mt: 0.5,
-                  p: 0,
-                  minWidth: 'auto',
-                  fontSize: { xs: '0.7rem', sm: '0.75rem' },
-                }}
-                endIcon={
-                  expandedDescriptions[group.groupId] ? <ExpandLessIcon /> : <ExpandMoreIcon />
-                }
-              >
-                {expandedDescriptions[group.groupId] ? 'Show less' : 'Show more'}
-              </Button>
-            )}
-          </Box>
-          <Box sx={{ display: 'flex', alignItems: 'center', mb: 1, flexWrap: 'wrap' }}>
-            <PeopleIcon fontSize="small" sx={{ mr: 0.5, color: 'text.secondary' }} />
-            <Typography
-              variant="body2"
-              color="text.secondary"
-              sx={{ fontSize: { xs: '0.8rem', sm: '0.875rem' } }}
-            >
-              {memberCount > 0 ? memberCount : '...'} {memberCount === 1 ? 'member' : 'members'}
-            </Typography>
-            <Chip
-              size="small"
-              label={getRoleLabel(group.userRole)}
-              sx={{
-                ml: 1,
-                height: { xs: 20, sm: 24 },
-                fontSize: { xs: '0.7rem', sm: '0.75rem' },
-              }}
-              color={getRoleColor(group.userRole)}
-            />
-          </Box>
-          <Chip
-            label={`Join Code: ${group.joinCode}`}
-            size="small"
-            variant="outlined"
-            sx={{
-              mt: 1,
-              height: { xs: 24, sm: 28 },
-              fontSize: { xs: '0.7rem', sm: '0.75rem' },
-            }}
-            onClick={e => {
-              e.stopPropagation(); // Prevent card click when clicking join code
-              handleCopyJoinCode(group.joinCode);
-            }}
-            icon={<ContentCopyIcon fontSize="small" />}
-            color={copiedCode === group.joinCode ? 'success' : 'default'}
-          />
-        </CardContent>
-        <CardActions sx={{ p: { xs: 1, sm: 2 }, pt: 0 }}>
-          {hasPermission(group, GroupPermission.MANAGE_MEMBERS) && !isDeletingThisGroup ? (
-            <Button
-              size="small"
-              color="primary"
-              variant="outlined"
-              startIcon={<ManageAccountsIcon />}
-              onClick={e => {
-                e.stopPropagation(); // Prevent card click when clicking manage members
-                handleOpenMembersDialog(group);
-              }}
-              sx={{
-                fontSize: { xs: '0.7rem', sm: '0.75rem' },
-                borderRadius: 2,
-                textTransform: 'none',
-                fontWeight: 500,
-                boxShadow: '0 1px 2px rgba(0,0,0,0.05)',
-                '&:hover': {
-                  backgroundColor: 'rgba(143, 197, 163, 0.08)',
-                  boxShadow: '0 2px 4px rgba(0,0,0,0.1)',
-                },
-              }}
-            >
-              Manage Members
-            </Button>
-          ) : null}
-        </CardActions>
-      </Card>
-    );
-  };
-
   if (!auth || !auth.firebaseUser || !auth.myUser) return null;
 
   if (loading) {
@@ -655,100 +385,20 @@ export default function Groups() {
         </Box>
       </Box>
 
-      {/* Search and Filter Section */}
-      <Paper
-        elevation={0}
-        sx={{
-          p: 2,
-          mb: 3,
-          borderRadius: 2,
-          border: '1px solid',
-          borderColor: 'divider',
-          backgroundColor: 'background.paper',
-        }}
-      >
-        <Box sx={{ display: 'flex', alignItems: 'center', mb: showFilters ? 2 : 0 }}>
-          <TextField
-            fullWidth
-            placeholder="Search groups by name, description, or join code"
-            value={searchQuery}
-            onChange={handleSearchChange}
-            variant="outlined"
-            size="small"
-            sx={{ mr: 2 }}
-            InputProps={{
-              startAdornment: (
-                <InputAdornment position="start">
-                  <SearchIcon color="action" />
-                </InputAdornment>
-              ),
-            }}
-          />
-          <Tooltip title={showFilters ? 'Hide filters' : 'Show filters'}>
-            <IconButton onClick={toggleFilters} color={showFilters ? 'primary' : 'default'}>
-              <FilterListIcon />
-            </IconButton>
-          </Tooltip>
-        </Box>
-
-        {/* Filter Options */}
-        {showFilters && (
-          <Box sx={{ mt: 2 }}>
-            <Box
-              sx={{ display: 'flex', flexDirection: { xs: 'column', sm: 'row' }, gap: 2, mb: 2 }}
-            >
-              <FormControl fullWidth size="small">
-                <InputLabel id="sort-label">Sort by</InputLabel>
-                <Select
-                  labelId="sort-label"
-                  value={sortOption}
-                  onChange={handleSortChange}
-                  label="Sort by"
-                >
-                  <MenuItem value="name">Name</MenuItem>
-                  <MenuItem value="members">Members</MenuItem>
-                  <MenuItem value="recent">Most Recent</MenuItem>
-                </Select>
-              </FormControl>
-            </Box>
-
-            <Tabs
-              value={activeFilterTab}
-              onChange={handleFilterTabChange}
-              variant="scrollable"
-              scrollButtons="auto"
-              aria-label="group filter tabs"
-              sx={{ borderBottom: 1, borderColor: 'divider' }}
-            >
-              <Tab label="All Groups" value="all" />
-              <Tab label="Admin" value="admin" icon={<ManageAccountsIcon />} iconPosition="start" />
-              <Tab label="Moderator" value="moderator" icon={<PeopleIcon />} iconPosition="start" />
-              <Tab label="Member" value="member" icon={<GroupIcon />} iconPosition="start" />
-            </Tabs>
-          </Box>
-        )}
-      </Paper>
-
-      {/* Filter Summary */}
-      {(searchQuery || activeFilterTab !== 'all') && (
-        <Box sx={{ mb: 2, display: 'flex', alignItems: 'center', flexWrap: 'wrap', gap: 1 }}>
-          <Typography variant="body2" color="text.secondary">
-            Showing {filteredGroups.length} of {groups.length} groups
-          </Typography>
-          {(searchQuery || activeFilterTab !== 'all') && (
-            <Button
-              size="small"
-              onClick={() => {
-                setSearchQuery('');
-                setActiveFilterTab('all');
-              }}
-              sx={{ ml: 1 }}
-            >
-              Clear filters
-            </Button>
-          )}
-        </Box>
-      )}
+      {/* Group Filters Component */}
+      <GroupFilters
+        searchQuery={searchQuery}
+        activeFilterTab={activeFilterTab}
+        sortOption={sortOption}
+        showFilters={showFilters}
+        groupsCount={groups.length}
+        filteredGroupsCount={filteredGroups.length}
+        onSearchChange={handleSearchChange}
+        onFilterTabChange={handleFilterTabChange}
+        onSortChange={handleSortChange as (event: SelectChangeEvent<SortOption>) => void}
+        onToggleFilters={toggleFilters}
+        onClearFilters={clearFilters}
+      />
 
       {filteredGroups.length === 0 ? (
         <Card sx={{ p: { xs: 2, sm: 4 }, textAlign: 'center' }}>
@@ -773,14 +423,7 @@ export default function Groups() {
               Create Your First Group
             </Button>
           ) : (
-            <Button
-              variant="outlined"
-              color="primary"
-              onClick={() => {
-                setSearchQuery('');
-                setActiveFilterTab('all');
-              }}
-            >
+            <Button variant="outlined" color="primary" onClick={clearFilters}>
               Clear Filters
             </Button>
           )}
@@ -793,7 +436,25 @@ export default function Groups() {
             gap: { xs: 2, sm: 3 },
           }}
         >
-          {filteredGroups.map(renderGroupCard)}
+          {filteredGroups.map(group => (
+            <GroupCard
+              key={group.groupId}
+              group={group}
+              memberCount={memberCounts[group.groupId] || 0}
+              isExpanded={!!expandedDescriptions[group.groupId]}
+              isDeleting={deletingGroupId === group.groupId}
+              copiedCode={copiedCode}
+              hasPermission={(group, permission) => hasPermission(group, permission)}
+              getRoleColor={role => getRoleColor(role as GroupMemberRole)}
+              getRoleLabel={role => getRoleLabel(role as GroupMemberRole)}
+              onNavigateToDetail={handleNavigateToGroupDetail}
+              onEdit={handleOpenEditDialog}
+              onDelete={handleOpenDeleteDialog}
+              onCopyJoinCode={handleCopyJoinCode}
+              onToggleDescription={toggleDescription}
+              onManageMembers={handleOpenMembersDialog}
+            />
+          ))}
         </Box>
       )}
